requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-bootstrap-v5==1.0.11
django-bootstrap4==22.3
django-bootstrap5==22.2
django-filter==23.1
docopt==0.6.2
idna==3.4
Pillow==9.4.0
pipreqs==0.4.11
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
yarg==0.1.9
